rename app.rs to app_manager.rs
