+++ riffle-server/src/http/apps.rs
@@ -1,4 +1,4 @@
-use crate::app::{self, App, APP_MANAGER_REF};
+use crate::app_manager::{self, App, APP_MANAGER_REF};
 use crate::http::{Format, Handler};
 use crate::util;
 use chrono::{Local, TimeZone, Utc};
