+++ riffle-server/src/metric.rs
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use crate::app::SHUFFLE_SERVER_ID;
+use crate::app_manager::SHUFFLE_SERVER_ID;
 use crate::config::Config;
 use crate::constant::CPU_ARCH;
 use crate::histogram;
